+++ back/src/main/java/com/ecommerce/vendamais/controller/ProductController.java
@@ -3,7 +3,6 @@ package com.ecommerce.vendamais.controller;
 import com.ecommerce.vendamais.dto.ProductDto;
 import com.ecommerce.vendamais.common.ApiResponse;
 import com.ecommerce.vendamais.model.Company;
-import com.ecommerce.vendamais.model.Photo;
 import com.ecommerce.vendamais.model.Type;
 import com.ecommerce.vendamais.repository.PhotoRepository;
 import com.ecommerce.vendamais.repository.TypeRepository;
@@ -61,13 +60,7 @@ public class ProductController {
 
     }
 
-    @GetMapping("/list/type/{typeId}")
-    public List<ProductDto> getAllProductsByTypeId(@PathVariable("typeId") int typeId){
-        return productService.getAllProductsByTypeId(typeId);
-
-    }
-
-    @PostMapping("/{productId}/uploadPhoto")
+    @PostMapping("/{productId}/upload")
     public ResponseEntity<ApiResponse> uploadPhoto(@RequestParam MultipartFile photo, int productId) throws Exception {
         if(!productService.findById(productId)){
             return new ResponseEntity<>(new ApiResponse(false, "produto não existe"), HttpStatus.NOT_FOUND);
@@ -76,14 +69,10 @@ public class ProductController {
         return new ResponseEntity<>(new ApiResponse(true, "foto do produto salva com sucesso"), HttpStatus.OK);
     }
 
-    @GetMapping("/{productId}/downloadPhoto")
+    @GetMapping("/{productId}/download")
     public @NotNull byte[] getProductPhotoById(@PathVariable("productId") int productId){
         return productService.getProductPhotoById(productId);
-    }
 
-    @GetMapping("/getProductPhotos")
-    public @NotNull List<Photo> getProductPhotos(){
-        return productService.getProductPhotos();
     }
