+++ back/src/main/java/com/ecommerce/vendamais/repository/ProductRepository.java
@@ -4,9 +4,6 @@ import com.ecommerce.vendamais.model.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
-
 @Repository
 public interface ProductRepository extends JpaRepository<Product, Integer> {
-    List<Product> findAllByTypeId(int typeId);
 }
