+++ web/src/api/produtos.json
@@ -7,15 +7,5 @@
         "preco": 1499,
         "precoAnterior": 1899,
         "tipoId": 5
-      },
-    
-      {
-        "descricao": "Uma potência no ar Com o chip M1 da Apple, nosso notebook mais fino e leve ficou irreconhecível. Por dentro. A CPU e a GPU estão mais rápidas, até 3,5 vezes e até cinco vezes, respectivamente. Nosso Neural Engine mais avançado deixa o aprendizado de máquina até nove vezes mais veloz. A bateria oferece mais tempo de duração. E trocamos a ventoinha pelo silêncio. Nunca tanto desempenho esteve reunido em um MacBook Air. Um pequeno chip, uma grande revolução Nosso primeiro chip projetado especialmente para o Mac acaba de chegar. Com impressionantes 16 bilhões de transistores, o M1 da Apple é um sistema em um chip (SoC). Isso quer dizer que ele reúne CPU, GPU, Neural Engine, conexões e outros recursos em uma peça só, e minúscula. Ele traz desempenho incrível, tecnologias exclusivas e o consumo de energia mais eficiente da indústria. O chip M1 é mais do que um avanço. É uma nova categoria. CPU de 8 núcleos Trabalha muito. A bateria, quase nada O M1 tem a CPU mais rápida que já produzimos. Com tanto processamento, o MacBook Air agora é capaz de realizar tarefas pesadas, como editar vídeos com qualidade profissional e rodar jogos cheios de ação. Além de ter velocidade até 3,5 vezes maior do que a geração anterior, a CPU de oito núcleos do chip M1 combina núcleos de alto desempenho e núcleos de eficiência para fazer trabalhos do dia a dia com apenas um décimo da energia.",
-        "estoque": 150,
-        "marca": " Apple",
-        "nome": "Macbook Air Apple 13.3, Processador M1, 8GB, SSD 256GB, Space Grey - MGN63BZ/A",
-        "preco": 6949.99,
-        "precoAnterior": 7315.78,
-        "tipoId": 1
-      }  
+      }
 ]
