--- web/src/interfaces/ListaTipoProduto.ts
@@ -1,5 +1,0 @@
-import { TipoProduto } from "./TipoProduto";
-
-export interface ListaTipoProduto{
-    listaTipoProduto: TipoProduto[]
-}
