+++ web/src/interfaces/TipoProduto.ts
@@ -1,4 +1,8 @@
+import { Categoria } from "./Categoria";
+
+
 export interface TipoProduto{
     id: number;
     nome: String;
+    categoria: Categoria
 }
